# docs/openapi-apifox.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "云舒 · 后端服务",
    "description": "AI 心理健康助手「云舒」后端服务。由 Java 版迁移而来，**接口契约与原版逐字对齐**。\n\n## 统一响应格式\n\n除 SSE 接口外，所有接口都返回同一个信封：\n\n```json\n{ \"code\": \"200\", \"msg\": \"success\", \"data\": {} }\n```\n\n- `code` 是**字符串**，`\"200\"` 表示成功，其余为业务错误码\n- **业务失败通常仍是 HTTP 200**，必须判 `code` 而不是 HTTP 状态码\n- 少数情况 HTTP 与 code 一起变：未登录 401、参数校验失败 400、越权 400\n\n## 鉴权\n\n登录 `/api/user/login` 拿到 `data.token`，之后所有需鉴权接口带：\n\n```\nAuthorization: Bearer <token>\n```\n\n**在 Apifox 里只需设置一次**：登录接口 → 后置操作 → 提取变量 `data.token`，\n然后把它绑到全局的 Bearer 认证即可（详见项目 README）。\n\n需鉴权的接口占 35/39。**无需鉴权的只有 4 个**：登录、注册、刷新令牌、登出。\n\n## 权限\n\n`userType`：`1` = 普通用户，`2` = 管理员。带 👑 的接口需要管理员，\n普通用户调用返回 **HTTP 400 + code `A0300`**（不是 403，这是原版的行为）。\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/crisis/page": {
      "get": {
        "tags": [
          "危机预警"
        ],
        "summary": "危机事件分页",
        "description": "⚠️ size 上限是 **50**（不是全局的 100），原版硬编码如此。\n\n👑 需要管理员（userType=2），普通用户调用返回 HTTP 400 + code A0300。",
        "operationId": "page_api_admin_crisis_page_get",
        "parameters": [
          {
            "name": "currentPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Currentpage"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 10,
              "title": "Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_dict_"
                },
                "example": {
                  "code": "200",
                  "msg": "操作成功",
                  "data": {
                    "records": [
                      {
                        "id": 12,
                        "userId": 24,
                        "sessionId": 55,
                        "diaryId": null,
                        "source": "CHAT",
                        "level": 3,
                        "triggerType": "KEYWORD",
                        "matchedTerms": "自杀",
                        "contentSnippet": "我今天想自杀咋办呀",
                        "status": "PENDING",
                        "handlerId": null,
                        "handleNote": null,
                        "handledAt": null,
                        "createdAt": "2026-10-02T14:34:30",
                        "updatedAt": "2026-10-02T14:34:30"
                      },
                      {
                        "id": 11,
                        "userId": 47,
                        "sessionId": null,
                        "diaryId": null,
                        "source": "SCALE",
                        "level": 3,
                        "triggerType": "SCALE",
                        "matchedTerms": "PHQ9-Q9",
                        "contentSnippet": "PHQ9 自伤项得分 3（总分 3）",
                        "status": "RESOLVED",
                        "handlerId": 48,
                        "handleNote": "E2E 测试处置",
                        "handledAt": "2026-10-01T19:43:31",
                        "createdAt": "2026-10-01T19:43:19",
                        "updatedAt": "2026-10-01T19:43:31"
                      },
                      {
                        "id": 10,
                        "userId": 45,
                        "sessionId": null,
                        "diaryId": null,
                        "source": "SCALE",
                        "level": 3,
                        "triggerType": "SCALE",
                        "matchedTerms": "PHQ9-Q9",
                        "contentSnippet": "PHQ9 自伤项得分 3（总分 3）",
                        "status": "RESOLVED",
                        "handlerId": 46,
                        "handleNote": "E2E 测试处置",
                        "handledAt": "2026-10-01T19:39:30",
                        "createdAt": "2026-10-01T19:39:20",
                        "updatedAt": "2026-10-01T19:39:30"
                      },
                      {
                        "id": 9,
                        "userId": 43,
                        "sessionId": null,
                        "diaryId": null,
                        "source": "SCALE",
                        "level": 3,
                        "triggerType": "SCALE",
                        "matchedTerms": "PHQ9-Q9",
                        "contentSnippet": "PHQ9 自伤项得分 3（总分 3）",
                        "status": "PENDING",
                        "handlerId": null,
                        "handleNote": null,
                        "handledAt": null,
                        "createdAt": "2026-10-01T19:38:02",
                        "updatedAt": "2026-10-01T19:38:02"
                      },
                      {
                        "id": 8,
                        "userId": 28,
                        "sessionId": 49,
                        "diaryId": null,
                        "source": "CHAT",
                        "level": 1,
                        "triggerType": "KEYWORD",
                        "matchedTerms": "焦虑",
                        "contentSnippet": "推荐几本关于焦虑的书",
                        "status": "PENDING",
                        "handlerId": null,
                        "handleNote": null,
                        "handledAt": null,
                        "createdAt": "2026-10-01T15:47:01",
                        "updatedAt": "2026-10-01T15:47:01"
                      },
                      {
                        "id": 7,
                        "userId": 28,
                        "sessionId": 46,
                        "diaryId": null,
                        "source": "CHAT",
                        "level": 3,
                        "triggerType": "KEYWORD",
                        "matchedTerms": "不想活,不想活了",
                        "contentSnippet": "我不想活了",
                        "status": "PENDING",
                        "handlerId": null,
                        "handleNote": null,
                        "handledAt": null,
                        "createdAt": "2026-10-01T15:46:22",
                        "updatedAt": "2026-10-01T15:46:22"
                      },
                      {
                        "id": 6,
                        "userId": 27,
                        "sessionId": 45,
                        "diaryId": null,
                        "source": "CHAT",
                        "level": 1,
                        "triggerType": "KEYWORD",
                        "matchedTerms": "焦虑",
                        "contentSnippet": "我考试前总是特别焦虑，有什么办法缓解吗",
                        "status": "PENDING",
                        "handlerId": null,
                        "handleNote": null,
                        "handledAt": null,
                        "createdAt": "2026-10-01T15:38:08",
                        "updatedAt": "2026-10-01T15:38:08"
                      },
                      {
                        "id": 5,
                        "userId": 26,
                        "sessionId": 44,
                        "diaryId": null,
                        "source": "CHAT",
                        "level": 1,
                        "triggerType": "KEYWORD",
                        "matchedTerms": "焦虑",
                        "contentSnippet": "我考试前总是特别焦虑，有什么办法缓解吗",
                        "status": "PENDING",
                        "handlerId": null,
                        "handleNote": null,
                        "handledAt": null,
                        "createdAt": "2026-10-01T15:37:17",
                        "updatedAt": "2026-10-01T15:37:17"
                      },
                      {
                        "id": 4,
                        "userId": 25,
                        "sessionId": null,
                        "diaryId": null,
                        "source": "SCALE",
                        "level": 3,
                        "triggerType": "SCALE",
                        "matchedTerms": "PHQ9-Q9",
                        "contentSnippet": "PHQ9 自伤项得分 3（总分 3）",
                        "status": "PENDING",
                        "handlerId": null,
                        "handleNote": null,
                        "handledAt": null,
                        "createdAt": "2026-10-01T15:19:53",
                        "updatedAt": "2026-10-01T15:19:53"
                      },
                      {
                        "id": 3,
                        "userId": 20,
                        "sessionId": 42,
                        "diaryId": null,
                        "source": "CHAT",
                        "level": 1,
                        "triggerType": "KEYWORD",
                        "matchedTerms": "压力好大",
                        "contentSnippet": "我最近压力好大",
                        "status": "PENDING",
                        "handlerId": null,
                        "handleNote": null,
                        "handledAt": null,
                        "createdAt": "2026-10-01T13:54:41",
                        "updatedAt": "2026-10-01T13:54:41"
                      }
                    ],
                    "total": 12
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/crisis/{event_id}/handle": {
      "post": {
        "tags": [
          "危机预警"
        ],
        "summary": "处置危机事件",
        "description": "status 为空时默认 RESOLVED；只有 PENDING/HANDLING 可处置。\n\n👑 需要管理员（userType=2），普通用户调用返回 HTTP 400 + code A0300。",
        "operationId": "handle_api_admin_crisis__event_id__handle_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrisisHandleIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/file/upload": {
      "post": {
        "tags": [
          "文件"
        ],
        "summary": "文件上传",
        "description": "multipart 表单：file + 可选的 businessType / businessId / businessField。",
        "operationId": "upload_api_file_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_file_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_FileOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_api_file_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "businessType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Businesstype"
          },
          "businessId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Businessid"
          },
          "businessField": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Businessfield"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_api_file_upload_post"
      },
...
      "CrisisHandleIn": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "RESOLVED",
                  "IGNORED",
                  "HANDLING"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "handleNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handlenote"
          }
        },
        "type": "object",
        "title": "CrisisHandleIn",
        "description": "status 为空时默认 RESOLVED，这是原版行为。"
      },
...
      "FileOut": {
        "properties": {
          "fileId": {
            "type": "integer",
            "title": "Fileid"
          },
          "filePath": {
            "type": "string",
            "title": "Filepath"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "originalName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Originalname"
          },
          "fileType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filetype"
          },
          "fileSize": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filesize"
          }
        },
        "type": "object",
        "required": [
          "fileId",
          "filePath",
          "url"
        ],
        "title": "FileOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Result_FileOut_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "业务码，字符串类型",
            "default": "200"
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "提示信息",
            "default": "操作成功"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据"
          }
        },
        "type": "object",
        "title": "Result[FileOut]"
      },
...
      "Result_NoneType_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "业务码，字符串类型",
            "default": "200"
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "提示信息",
            "default": "操作成功"
          },
          "data": {
            "type": "null",
            "title": "Data",
            "description": "业务数据"
          }
        },
        "type": "object",
        "title": "Result[NoneType]"
      },
...
      "Result_dict_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "业务码，字符串类型",
            "default": "200"
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "提示信息",
            "default": "操作成功"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "业务数据"
          }
        },
        "type": "object",
        "title": "Result[dict]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录接口返回的 data.token。在 Apifox 里**只填 token 本身**，会自动补 `Bearer ` 前缀。也可填 `{{token}}` 引用全局变量。"
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:1236",
      "description": "本地直连后端（推荐）"
    },
    {
      "url": "http://127.0.0.1:5173",
      "description": "经前端 Vite 代理（需前端在运行）"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
}